requirements.txt
aiocoap
flask
influxdb-client
celery
redis

